# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Silence Notes API
  description: |
    ## Overview

    Silence Notes is a Chrome extension for note-taking with hashtag filtering and persistent backend storage.
    This API provides authentication, user management, and note functionality with comprehensive security features.

    ## Authentication

    The API uses Google OAuth 2.0 for authentication with JWT tokens. Protected endpoints require a valid Bearer token:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Security Features

    - **Rate Limiting**: Global (100 req/sec) and per-user (60 req/min) limits
    - **CORS Protection**: Configured for Chrome extension and web origins
    - **Security Headers**: CSP, HSTS, XSS protection, frame options
    - **Session Management**: Concurrent session limits and activity tracking
    - **Request Validation**: Input sanitization and size limits

    ## Error Handling

    All errors return JSON with consistent format:
    ```json
    {
      "error": "Error message",
      "code": 400,
      "success": false
    }
    ```

    ## Rate Limiting

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Time when rate limit window resets

  version: 1.0.0
  contact:
    name: Silence Notes Team
    email: support@silenzenotes.com
  license:
    name: MIT
...
servers:
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Uses a refresh token to obtain a new access token. Refresh tokens are
        long-lived and can be used to maintain user sessions without requiring
        re-authentication.
      operationId: refreshToken
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
              example:
                success: true
                access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                token_type: Bearer
                expires_in: 3600
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Invalid or expired refresh token"
                code: 401
                success: false
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token obtained from authentication flow"
  schemas:
...
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Valid refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    TokenRefreshResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        access_token:
          type: string
          description: New JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: New access token expiration time in seconds
          example: 3600
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request"
        code:
          type: integer
          description: HTTP status code
          example: 400
        success:
          type: boolean
          example: false
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Invalid request data"
            code: 400
            success: false
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal server error"
            code: 500
            success: false
...